requirements.txt
asgiref==3.5.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.12.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.2.0
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
